add Validation

+++ Controller/category.controller.js
@@ -1,5 +1,6 @@
 let Category = require('../model/Category');
 let sequelizeCon = require('../config/database.config');
+const { Error } = require('sequelize');
 
 async function createTable() {
   await sequelizeCon.sync({ force: true });
@@ -22,18 +23,33 @@ async function insertCategoies() {
 //---------------C  => Create---------------------
 // adding multiple Categories
 let insertCategories = async (req, res, next) => {
-  let body = req.body;
-  await Category.bulkCreate(body);
-  res.status(203).send('Categories Added Successfully');
-  res.end();
+  try {
+    let body = req.body;
+    if (body == ' ') {
+      throw new Error();
+    }
+    await Category.bulkCreate(body);
+    res.status(203).send('Categories Added Successfully');
+    res.end();
+  } catch (err) {
+    res.status(400).send('Error : Enter Right Value');
+    res.end();
+  }
 };
 
 // adding single category
 let insertCategory = async (req, res, next) => {
-  let body = req.body;
-  await Category.create(body);
-  res.status(203).send('Category added Successfully');
-  res.end();
+  try {
+    let body = req.body;
+    if (body.name == ' ') {
+      throw new Error();
+    }
+    await Category.create(body);
+    res.status(203).send('Category added Successfully');
+    res.end();
+  } catch (err) {
+    res.status(400).send('Error : Enter Right Value');
+  }
 };
 //--------------------------------------------------------------
 
@@ -48,37 +64,65 @@ let getAllCategories = async (req, res, next) => {
   res.end();
 };
 let getCategoryById = async (req, res, next) => {
-  let id = req.params.id;
-  let cat = await Category.findAll({
-    where: {
-      id: id,
-    },
+  let lastId = await Category.findOne({
+    order: [['id', 'DESC']],
   });
-  res.status(201).send(JSON.stringify(cat, null, 2));
-  res.end();
+  try {
+    let id = req.params.id;
+    let cat = await Category.findAll({
+      where: {
+        id: id,
+      },
+    });
+    if (id > lastId.id) {
+      throw new Error('Not Fount');
+    } else {
+      console.log(typeof cat);
+      res.status(201).send(JSON.stringify(cat, null, 2));
+      res.end();
+    }
+  } catch (err) {
+    res.status(400).send('Error Category not Found');
+    res.end();
+  }
 };
 
 //----------------U => Update ----------------------------------
 
 let updateCategories = async (req, res, next) => {
-  let body = req.body;
-  await Category.update(
-    {
-      name: body.name,
-    },
-    {
-      where: {
-        id: body.id,
-      },
+  try {
+    let body = req.body;
+    if (body.name == '' || body.id == NaN) {
+      throw new Error();
     }
-  );
-  res.status(202).send('Category was Updated');
-  res.end();
+    await Category.update(
+      {
+        name: body.name,
+      },
+      {
+        where: {
+          id: body.id,
+        },
+      }
+    );
+    res.status(202).send('Category was Updated');
+    res.end();
+  } catch (err) {
+    res.status(202).send('Error Something is Missing in JSON ');
+    res.end();
+  }
 };
 
 //--------------- D => Delete -----------------------------------
 let deleteCategories = async (req, res, next) => {
+    try{
+   let lastId = await Category.findOne({
+      order: [['id', 'DESC']],
+    });
   let id = req.params.id;
+  if(id>lastId.id){
+    throw new Error()
+  }
   await Category.destroy({
     where: {
       id: id,
@@ -86,6 +130,10 @@ let deleteCategories = async (req, res, next) => {
   });
   res.status(202).send('Delete Successfully');
   res.end();
+} catch(err){
+    res.status(400).send("Error : This Id is not in list")
+    res.end()
+}
 };
 
 module.exports = {
